requirements.txt
torch
torchvision
pillow
tqdm
matplotlib
gradio
opencv-python
numpy
